baselines/requirements.txt
hydra-core
jinja2
lmstudio
omegaconf
openai
pandas
pydantic
PyYAML
tqdm
scikit-learn
seaborn
matplotlib
numpy
scipy
